Rename quantized inner product GEMM V0
