+++ docs/2026.html
@@ -196,6 +196,7 @@
 <h5>Renaming</h5>
 <ul>
  <li>Class SynetQuantizedConvolutionNhwcGemm to SynetQuantizedConvolutionNhwcGemmV0.</li>
+ <li>Class SynetQuantizedInnerProductGemmNN to SynetQuantizedInnerProductGemmV0.</li>
 </ul>
 
 <h4>Test framework</h4>
