+++ prj/vs2022/AmxBf16.vcxproj.filters
@@ -299,7 +299,7 @@
     <ClCompile Include="..\..\src\Simd\SimdAmxBf16SynetQuantizedInnerProduct.cpp">
       <Filter>AmxBf16\Synet\Quantized</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Simd\SimdAmxBf16SynetQuantizedInnerProductGemmNN.cpp">
+    <ClCompile Include="..\..\src\Simd\SimdAmxBf16SynetQuantizedInnerProductGemmV0.cpp">
       <Filter>AmxBf16\Synet\Quantized</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\Simd\SimdAmxBf16SynetQuantizedMergedConvolution.cpp">
